+++ config/general/networkConfig.js
@@ -6,4 +6,4 @@ let devnet = {
     network: 'https://net.ton.dev'
 }
 
-module.exports = devnet;
+module.exports = tonOSSE;
